fix some compiler warnings

+++ src/gui/src/IpcClient.cpp
@@ -106,7 +106,7 @@ void IpcClient::sendCommand(const QString& command, ElevateMode const elevate)
 
     std::string stdStringCommand = command.toStdString();
     const char* charCommand = stdStringCommand.c_str();
-    int length = strlen(charCommand);
+    int length = (int)strlen(charCommand);
 
     char lenBuf[4];
     intToBytes(length, lenBuf, 4);
